Fleshed out the Previous Sales component. This will display all the sales record, from any sales rep (I'm assuming this is the requirement).
I'm using Livewire events to trigger a refresh of the Previous Sales component when a new sales is created. This is implemented by getting the NewSaleForm component to emit the event after creating the sale, and having PreviousSales listen on it.

If there are no sales, the message "No sale yet" will be displayed.