Added ShippingChargeSeeder and ShippingChargeFactory.
There may be lots of sales already using the existing shipping charge, so we seed the shipping_charges table with this first charge. The number of sales that uses this rate will be all the sales to date, so we set the number of sales of this first record to the total number of sales.

Also updated the DatabaseSeeder to run our new ShippingChargeSeeder.